move visualscript components to panels
