+++ packages/editor/src/panels/visualscript/hooks/useTemplateHandler.ts
