+++ packages/editor/src/panels/visualscript/util/calculateNewEdge.ts
