+++ packages/editor/src/panels/visualscript/util/hasPositionMetaData.ts
